ref #10525: adjusted rate for author comparison

+++ cdmlib-persistence/src/main/java/eu/etaxonomy/cdm/persistence/dao/name/INameMatchingDao.java
@@ -40,5 +40,5 @@ public interface INameMatchingDao extends IIdentifiableDao<TaxonName> {
      *     Names to be excluded from the result set
 	 * @return
 	 */
-    public List<NameMatchingParts> findNameMatchingParts(Map<String, Integer> map);
+    public List<NameMatchingParts> findNameMatchingParts(Map<String, Double> map);
 }
